requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-debug-toolbar==1.11
django-invitations==1.9.2
idna==2.8
oauthlib==2.1.0
PyJWT==1.7.1
python-constraint==1.4.0
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
rules==2.0.1
six==1.12.0
sqlparse==0.2.4
urllib3==1.24.1
